feat: Add `passthrough` option (#12)
* feat: Add passthrough ability

This allows a user to use the exec-action to only set the Percy env vars that
are needed for things to work correctly. This also adds a few env var tests

* chore: Rename CI workflow

* fix: Handle GitHub payloads better

* ci: Split PRs & push workflows

This is because you _have_ to use a JS action (it seems, would love to be wrong)
to gather the branch name on PR workflows. So we'll rely on the Percy build
being correct in order to test this

* Unclosure `setPercyBranchBuildInfo`